+++ src/main/java/ru/skypro/homework/repositories/UserRepository.java
@@ -5,9 +5,9 @@ import ru.skypro.homework.models.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
 
     Optional<User> findUserByEmailIs(String email);
